Move Avatars and TeamAvatar into their own file
